+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/AbstractBoxTest.kt
@@ -98,6 +98,7 @@ open class AbstractIrOnlyClassesBoxTest : AbstractBoxTest() {
     with(builder) {
       defaultDirectives {
         MetroDirectives.MIN_COMPILER_VERSION.with("2.4.20-dev-6138")
+        +MetroDirectives.GENERATE_CONTRIBUTION_HINTS_IN_FIR
         MetroDirectives.GENERATE_CLASSES_IN_IR.with(true)
       }
     }